FE/Qt4: New running VM core: win build fix #3.
git-svn-id: http://www.virtualbox.org/svn/vbox/trunk@26998 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Frontends/VirtualBox/src/runtime/UIMachineView.h
@@ -206,7 +206,7 @@ private:
 #endif
 
     /* Private helpers: */
-    void fixModifierState(int *piCodes, uint *puCount);
+    void fixModifierState(LONG *piCodes, uint *puCount);
     QPoint viewportToContents(const QPoint &vp) const;
     void scrollBy(int dx, int dy);
 #ifdef VBOX_WITH_VIDEOHWACCEL
